requirements.txt
requests>=2.31.0
dbus-python>=1.3.0
